Convert a ws client test
